refactor: remove v1 MetadataPopover component
