+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/v2/Details/getExecutionDuration.tsx
