+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/v2/Details/index.tsx
@@ -13,8 +13,8 @@ import {Link} from 'modules/components/Link';
 import {Paths} from 'modules/Routes';
 import {tracking} from 'modules/tracking';
 import {Header} from '../../Header';
-import {SummaryDataKey, SummaryDataValue} from '../../styled';
-import {getExecutionDuration} from '../../Details/getExecutionDuration';
+import {SummaryDataKey, SummaryDataValue} from '../styled';
+import {getExecutionDuration} from './getExecutionDuration';
 import {type V2MetaDataDto} from '../types';
 import type {BusinessObject} from 'bpmn-js/lib/NavigatedViewer';
 import {DetailsModal} from './DetailsModal';
