+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/v2/Incidents/multiIncidents.tsx
@@ -7,7 +7,7 @@
  */
 
 import {Header} from '../../Header';
-import {SummaryDataValue} from '../../styled';
+import {SummaryDataValue} from '../styled';
 
 type Props = {count: number; onButtonClick: () => void};
