+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/v2/index.tsx
@@ -6,7 +6,7 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {Popover, Content, Divider} from '../styled';
+import {Popover, Content, Divider} from './styled';
 import {flowNodeMetaDataStore} from 'modules/stores/flowNodeMetaData';
 import {flowNodeSelectionStore} from 'modules/stores/flowNodeSelection';
 import {observer} from 'mobx-react';
